Development

+++ app/Http/Controllers/Admin/TripRouteController.php
@@ -9,6 +9,7 @@ use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
 use App\Exports\TripRoutesExport;
 use App\Imports\TripRoutesImport;
+use Illuminate\Support\Facades\Gate;
 
 class TripRouteController extends Controller
 {
@@ -17,6 +18,7 @@ class TripRouteController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index_vehicles_vehicle_route');
         $routes = TripRoute::with('category')->latest()->get();
 
         return view('layouts.admin.trip_routes.index', compact('routes'));
@@ -25,6 +27,7 @@ class TripRouteController extends Controller
 
     public function create()
     {
+        Gate::authorize('create_vehicles_vehicle_route');
         $categories = TripCategory::pluck('name', 'id');
 
         return view('layouts.admin.trip_routes.create', compact('categories'));
@@ -33,6 +36,7 @@ class TripRouteController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create_vehicles_vehicle_route');
         TripRoute::create($request->all());
 
         return redirect()->route('admin.trip-routes.index')
@@ -41,6 +45,7 @@ class TripRouteController extends Controller
 
     public function show()
     {
+        Gate::authorize('read_vehicles_vehicle_route');
         $categories = TripCategory::with('routes')->get();
 
         return view('layouts.admin.trip_routes.show', compact('categories'));
@@ -48,11 +53,13 @@ class TripRouteController extends Controller
 
     public function categoryView()
     {
+        Gate::authorize('read_vehicles_vehicle_route');
         $categories = TripCategory::with('routes')->get();
         return view('layouts.admin.trip_routes.show', compact('categories'));
     }
     public function edit($id)
     {
+        Gate::authorize('update_vehicles_vehicle_route');
         $route = TripRoute::findOrFail($id);
 
         $categories = TripCategory::pluck('name', 'id');
@@ -66,6 +73,7 @@ class TripRouteController extends Controller
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('update_vehicles_vehicle_route');
         $route = TripRoute::findOrFail($id);
 
         $route->update($request->all());
@@ -77,6 +85,7 @@ class TripRouteController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('delete_vehicles_vehicle_route');
         TripRoute::destroy($id);
 
         return back()->with('success', 'Deleted successfully');
